Added datetsamp and fixed some bugs

+++ RemoveAllContainers.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-docker rm -vf $(docker ps -aq)
+if [[ -n "$(docker ps -q)" ]]; then
+    docker rm -vf $(docker ps -aq)
+else
+    echo "No containers to remove."
+fi
